✨ Add platformer physics
- Player can move left and right
- Player can jump

+++ gd4_sfml_tcp_game/Utility.cpp
@@ -181,7 +181,7 @@ int Utility::Length(sf::Vector2f vector)
 }
 
 
-int Utility::sign(int num) {
+int Utility::sign(float num) {
 	return (num > 0) ? 1 : ((num < 0) ? -1 : 0);
 }
