+++ gd4_sfml_tcp_game/Utility.hpp
@@ -16,6 +16,6 @@ class Utility
 		static double ToDegrees(double angle);
 		static int RandomInt(int exclusive_max);
 		static int Length(sf::Vector2f vector);
-		static int sign(int num);
+		static int sign(float num);
 };
